+++ src/Common/ApplicationEvents.fs
@@ -1,4 +1,4 @@
-namespace KafkaApplication
+namespace Lmc.KafkaApplication
 
 module internal ApplicationEvents =
     open System