+++ src/Common/Types.fs
@@ -1,4 +1,4 @@
-namespace KafkaApplication
+namespace Lmc.KafkaApplication
 
 open Kafka
 open Metrics